+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `weight` option for LightGBM and linear regression
 - Added `intercept` option for linear regression
+- Fixed `Unknown label` error for LightGBM
 
 ## 0.3.0 (2019-09-05)